React hook form and zod validation

+++ src/actions/actions.ts
@@ -1,17 +1,23 @@
 'use server';
 
 import prisma from '@/lib/db';
-import { PetEssential } from '@/lib/types';
 import { sleep } from '@/lib/utils';
-import { Pet } from '@prisma/client';
+import { petFormSchema, petIdSchema } from '@/lib/validations';
 import { revalidatePath } from 'next/cache';
 
-export async function addPet(petData: PetEssential) {
+export async function addPet(petData: unknown) {
   await sleep(1000);
+
+  const validatedPetData = petFormSchema.safeParse(petData);
+  if (!validatedPetData.success) {
+    return {
+      message: 'Invalid pet data',
+    };
+  }
 
   try {
     await prisma.pet.create({
-      data: petData,
+      data: validatedPetData.data,
     });
   } catch (error) {
     return {
@@ -22,15 +28,24 @@ export async function addPet(petData: PetEssential) {
   revalidatePath('/app', 'layout');
 }
 
-export async function editPet(petId: Pet['id'], petData: PetEssential) {
+export async function editPet(petId: unknown, petData: unknown) {
   await sleep(1000);
+
+  const validatedPetId = petIdSchema.safeParse(petId);
+  const validatedPetData = petFormSchema.safeParse(petData);
+
+  if (!validatedPetData.success || !validatedPetId.success) {
+    return {
+      message: 'Invalid pet data',
+    };
+  }
 
   try {
     await prisma.pet.update({
       where: {
-        id: petId,
+        id: validatedPetId.data,
       },
-      data: petData,
+      data: validatedPetData.data,
     });
   } catch (error) {
     return {
@@ -41,13 +56,21 @@ export async function editPet(petId: Pet['id'], petData: PetEssential) {
   revalidatePath('/app', 'layout');
 }
 
-export async function deletePet(petId: Pet['id']) {
+export async function deletePet(petId: unknown) {
   await sleep(1000);
+
+  const validatedPetId = petIdSchema.safeParse(petId);
+
+  if (!validatedPetId.success) {
+    return {
+      message: 'Invalid pet data',
+    };
+  }
 
   try {
     await prisma.pet.delete({
       where: {
-        id: petId,
+        id: validatedPetId.data,
       },
     });
   } catch (error) {
